requirements.txt
aiofiles==23.1.0
aiomysql==0.2.0
annotated-types==0.5.0
arq==0.25.0
async-timeout==4.0.2
cffi==1.15.1
click==8.1.6
cryptography==41.0.2
greenlet==2.0.2
hiredis==2.2.3
html5tagger==1.3.0
httptools==0.6.0
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.5
multidict==6.0.4
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
PyMySQL==1.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
redis==4.6.0
sanic==23.6.0
sanic-ext==23.3.0
sanic-routing==23.6.0
sanic-session==0.8.0
SQLAlchemy==2.0.19
tracerite==1.1.0
typing_extensions==4.7.1
ujson==5.8.0
uvloop==0.17.0
websockets==11.0.3
